point 1 2 3

+++ cd2024bfs1g1-frontend/src/main/ngx/src/app/public/cart/cart-item/cart-item.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input, Injector, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Input, Injector, Output, EventEmitter, ViewChild, ElementRef } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { OntimizeService } from 'ontimize-web-ngx';
 import { CartService } from 'src/app/shared/services/cart.service';
@@ -14,13 +14,15 @@ export class CartItemComponent implements OnInit {
   @Input() showBtns: boolean = true;
   @Output() updateCart = new EventEmitter<void>();
   service: OntimizeService;
-  product: any={};  
-
+  product: any = {};
+  @ViewChild('less') less :ElementRef ;
+  isButtonDisabled;
 
   constructor(
     protected injector: Injector,
     protected sanitizer: DomSanitizer,
     private cartService: CartService
+    
   ) {
     this.service = this.injector.get(OntimizeService)
   }
@@ -35,19 +37,22 @@ export class CartItemComponent implements OnInit {
           this.itemAmount(this.product)
         }
       })
+      if (this.item.units <= 1) {
+        this.isButtonDisabled =true;
+      }
   }
 
   public getImageSrc(base64: any): any {
     return base64 ? this.sanitizer.bypassSecurityTrustResourceUrl('data:image/*;base64,' + base64) : './assets/images/no-image.png';
   }
-/*
-  get price() {
-    return this.product.PRO_PRICE?.toFixed(2);
-  }
-  get sale() {
-    return this.product.PRO_SALE?.toFixed(2);
-  }
-  */
+  /*
+    get price() {
+      return this.product.PRO_PRICE?.toFixed(2);
+    }
+    get sale() {
+      return this.product.PRO_SALE?.toFixed(2);
+    }
+    */
   public itemAmount(product: any): any {
     return product.PRO_PRICE * this.item.units;
   }
@@ -55,17 +60,25 @@ export class CartItemComponent implements OnInit {
 
   public addItem() {
     this.cartService.addProductToCart(this.product);
-    this.item.units = this.cartService.getCart().find(x=>x.id == this.item.id).units;
+    this.item.units = this.cartService.getCart().find(x => x.id == this.item.id).units;
     this.updateCart.emit();
+    if(this.item.units >1){
+      (<HTMLButtonElement> this.less.nativeElement).disabled =false;
+    }    
   }
 
   public removeItem() {
-    this.cartService.removeItem(this.product.PRO_ID);
-    let p_ls =this.cartService.getCart().find(x=>x.id == this.item.id);
-    if(p_ls !== undefined){
-      this.item.units = p_ls.units;
+    if (this.item.units > 1) {
+      this.cartService.removeItem(this.product.PRO_ID);
+      let p_ls = this.cartService.getCart().find(x => x.id == this.item.id);
+      if (p_ls !== undefined) {
+        this.item.units = p_ls.units;
+      }
+      this.updateCart.emit();
     }
-    this.updateCart.emit();
+    if (this.item.units <= 1) {
+      (<HTMLButtonElement> this.less.nativeElement).disabled =true;
+    }
   }
 
   public deleteItem() {
